+++ CHANGES.md
@@ -1,3 +1,8 @@
+## master
+
+* fix and add tests for `drop_labels` and `drop_labels_exn`, where the semantics
+  of the labeled `back` argument was inversed.
+
 ## 0.1.0 (2018-06-26)
 
 * Initial release